requirements.txt
PyYAML
FastAPI
Uvicorn
httpx
